added makefile for easy building & made license more permissive
Now in order to build the server all you need to do is write 'make' and
it downloads & builds bam and builds the server executable as well.
The Makefile usually only works on Linux and macOS

The license was made more permissive to allow normal clans to be easily
able to host a server, if they want to.

Added some (hopefully) easily to understand instructions on how to create
everything.